# swagger-spec.json
{"openapi":"3.0.0","paths":{"/v1/auth/registration":{"post":{"operationId":"AuthController_registration","summary":"Регистрация пользователя","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationDto"}}}},"responses":{"200":{"description":"Пользователь успешно создан.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorizedDto"}}}},"409":{"description":"Данный почтовый адрес уже занят."},"500":{"description":"Запрос закончился исключением"}},"tags":["Authorization"]}},"/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"Авторизация пользователя","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Успешно авторизован.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorizedDto"}}}},"401":{"description":"Пользователь с такими данными не был найден."},"500":{"description":"Запрос закончился исключением"}},"tags":["Authorization"]}},"/v1/auth/exit":{"post":{"operationId":"AuthController_exit","summary":"Выход пользователя","parameters":[],"responses":{"200":{"description":"Выход успешно выполнен."},"401":{"description":"Пользователь не авторизован."},"500":{"description":"Запрос закончился исключением"}},"tags":["Authorization"],"security":[{"Api-Key":[]}]}},"/v1/users/{id}/edit":{"put":{"operationId":"UsersController_edit","summary":"Редактирование профиля пользователя","parameters":[{"name":"id","required":true,"in":"path","description":"Индивидуальный идентификатор пользователя","schema":{}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditDto"}}}},"responses":{"200":{"description":"Информация успешно обновлена"},"401":{"description":"Нет доступа"},"404":{"description":"Пользователь не найден"},"500":{"description":"Запрос закончился исключением"}},"tags":["Users"],"security":[{"Api-Key":[]}]}},"/v1/users/{id}/photo":{"post":{"operationId":"UsersController_updatePhoto","summary":"Обновление аватара пользователя","parameters":[{"name":"id","required":true,"in":"path","description":"Индивидуальный идентификатор пользователя","schema":{}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Аватар для пользователя успешно обновлён"},"401":{"description":"Нет доступа"},"404":{"description":"Пользователь не найден"},"500":{"description":"Запрос закончился исключением"}},"tags":["Users"],"security":[{"Api-Key":[]}]}},"/v1/users/{id}/banner":{"post":{"operationId":"UsersController_updateBanner","summary":"Обновление баннера пользователя","parameters":[{"name":"id","required":true,"in":"path","description":"Индивидуальный идентификатор пользователя","schema":{}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Баннер для пользователя успешно обновлён"},"401":{"description":"Нет доступа"},"404":{"description":"Пользователь не найден"},"500":{"description":"Запрос закончился исключением"}},"tags":["Users"],"security":[{"Api-Key":[]}]}},"/v1/users/profile/{urn}":{"get":{"operationId":"UsersController_getProfile","summary":"Получение профиля пользователя по его персональной urn","parameters":[{"name":"urn","required":true,"in":"path","description":"Персональный URN пользователя","schema":{}}],"responses":{"200":{"description":"Информация о пользователе успешно получена","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDto"}}}},"401":{"description":"Нет доступа"},"404":{"description":"Пользователь не найден"},"500":{"description":"Запрос закончился исключением"}},"tags":["Users"],"security":[{"Api-Key":[]}]}},"/v1/users":{"get":{"operationId":"UsersController_getUsers","summary":"Получение списка пользователей","parameters":[],"responses":{"200":{"description":"Информация о пользователях успешно получена","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}},"401":{"description":"Нет доступа"},"500":{"description":"Запрос закончился исключением"}},"tags":["Users"],"security":[{"Api-Key":[]}]}},"/v1/images/{imageKey}":{"get":{"operationId":"ImagesController_getImage","summary":"Получение картинки по ключу","parameters":[{"name":"imageKey","required":true,"in":"path","description":"Ключ фото","schema":{"type":"string"}}],"responses":{"200":{"description":"Информация успешно обновлена","content":{"image/*":{"schema":{"type":"string","format":"binary"}}}},"401":{"description":"Нет доступа"},"404":{"description":"Картинка не найдена"},"500":{"description":"Запрос закончился исключением"}},"tags":["Images"],"security":[{"Api-Key":[]}]}}},"info":{"title":"Yoldi Agency API","description":"API к тестовому заданию от Yoldi Agency. Telegram: @m1fov","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"Api-Key":{"type":"apiKey","in":"header","name":"authorization"}},"schemas":{"RegistrationDto":{"type":"object","properties":{"name":{"type":"string","description":"Имя пользователя","minLength":1,"maxLength":20},"email":{"type":"string","description":"Электронный почтовый адрес","minLength":5},"password":{"type":"string","description":"Пароль","minLength":6,"maxLength":20}},"required":["name","email","password"]},"AuthorizedDto":{"type":"object","properties":{"id":{"type":"number","description":"Индивидуальный идентификатор пользователя"},"token":{"type":"string","description":"API статический токен"}},"required":["id","token"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"Электронный почтовый адрес","minLength":5},"password":{"type":"string","description":"Пароль","minLength":5}},"required":["email","password"]},"EditDto":{"type":"object","properties":{"name":{"type":"string","description":"Имя пользователя","minLength":1,"maxLength":20},"aboutMe":{"type":"string","description":"Описание профиля","minLength":0,"maxLength":255},"urn":{"type":"string","description":"URN пользователя","minLength":0,"maxLength":10}},"required":["name","aboutMe","urn"]},"ProfileDto":{"type":"object","properties":{"name":{"type":"string","description":"Имя пользователя"},"aboutMe":{"type":"string","description":"Описание профиля"},"email":{"type":"string","description":"Электронный почтовый адрес"},"photoKey":{"type":"string","description":"Ключ фото профиля"},"bannerKey":{"type":"string","description":"Ключ баннера профиля"}},"required":["name","aboutMe","email","photoKey","bannerKey"]},"UserDto":{"type":"object","properties":{"name":{"type":"string","description":"Имя пользователя"},"photoKey":{"type":"string","description":"Ключ фото профиля"},"email":{"type":"string","description":"Электронный почтовый адрес"},"urn":{"type":"string","description":"Персональный URN пользователя"}},"required":["name","photoKey","email","urn"]}}}}